In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,20 +9,20 @@
 
 int main(void)
 {
-	int i, c, d, number;
+	long long int i, c, d, number;
 
 	c = 1;
 	d = 2;
 
-	printf("%d, %d", c, d);
+	printf("%lld, %lld", c, d);
 
 	for (i = 1; i < 25; i++)
 	{
 		number = c + d;
-		printf(", %d", number);
+		printf(", %lld", number);
 		c = number;
 		number = c + d;
-		printf(", %d", number);
+		printf(", %lld", number);
 		d = number;
 	}
 	printf("\n");
